+++ R/cosewic_tools.R
@@ -10,12 +10,13 @@
 #' section in 'Instructions for preparing COSEWIC status reports'
 #' for more details.
 #'
-#' By default the EOO is calculated only using the inner 95% of points (based on
-#' distance to the centroid). This is to ensure that a first-pass of the EOO
-#' does not reject a species from consideration if there are any outlier
-#' observations. However, for a final COSEWIC assessment report, it is likely
+#' By default the EOO is calculated using all points (`eoo_p = 1` for 100% of
+#' However, if you're working on rough data or want to do a rough first pass,
+#' you may wish to use `eoo_p = 0.95` or 95% of points (based on distance to the
+#' centroid). This will ensure outlier observations will not artificially
+#' inflate the EOO. However, for a final COSEWIC assessment report, it is likely
 #' better to carefully explore the data to ensure there are no outliers and then
-#' use the full data set (i.e. set `eoo_p = 1`).
+#' use the full data set (i.e. use the default of `eoo_p = 1`).
 #'
 #' The IAO is calculated by first assessing large grids (10x large than the
 #' specified size). Only then are smaller grids created within large grid cells
@@ -42,9 +43,9 @@
 #'   IAO. Default is COSEWIC requirement (2km, meaning 2x2km grids of 4km2).
 #'   Use caution if changing.
 #' @param eoo_p Numeric. The percentile to calculate the convex hull over.
-#'   Defaults to 0.95 for a 95% convex hull to ensure outlier points do not
-#'   artificially inflate the EOO. Note that for a final COSEWIC report, this
-#'   may not be appropriate. Set to 1 to include all points.
+#'   Defaults to 1 for a 100% convex hull to match COSEWIC requirements
+#'   (includes all points). Note that you may wish to use 0.95 for a 95% convex
+#'   hull to  ensure outlier points do not artificially inflate the EOO.
 #' @param eoo_clip sf (Multi)Polygon. A spatial object to clip the EOO to. May
 #'   be relevant when calculating EOOs for complex regions (i.e. long curved
 #'   areas) to avoid including area which cannot have observations.
@@ -73,7 +74,7 @@
 #'   - `n_occupied` - Number of IAO cells with at least one record
 #'   - `iao` - IAO value (`grid_size_km`^2 * `n_occupied`)
 #'   - `eoo_pXX` - EOO area calculated with a convex hull at percentile `eoo_p`
-#'     (e.g., 95%)
+#'     (e.g., 100% or 95%)
 #'
 #' `spatial` contains spatial data frames
 #' - `iao_sf` - Polygons of the IAO grids with the `n_records` per cell
@@ -81,7 +82,6 @@
 #'
 #'
 #' @examples
-#'
 #' # Using the included, test data on black-capped chickadees
 #'
 #' r <- cosewic_ranges(bcch)
@@ -137,7 +137,7 @@ cosewic_ranges <- function(
   species = "species_id",
   iao_grid_size_km = 2,
   iao_grid = NULL,
-  eoo_p = 0.95,
+  eoo_p = 1,
   eoo_clip = NULL,
   crs = "ESRI:102001",
   which = c("eoo", "iao"),
@@ -148,6 +148,16 @@ cosewic_ranges <- function(
   have_pkg_check("sf")
   df <- df_db_check(df_db)
   which_check(which)
+
+  # Alerts
+  rlang::inform(
+    paste0(
+      "As of naturecounts v0.5.0 `cosewic_ranges()` now uses a default of ",
+      "`eoo_p = 1` instead of `eoo_p = 0.95`."
+    ),
+    .frequency = "once",
+    .frequency_id = "eoo_p"
+  )
 
   # Coords
   if (!all(c(coord_lat, coord_lon) %in% names(df))) {